Bump yahoo-finance-rs, and improve parallelization on position calc
The released version of yahoo-finance-rs crate has a few issues that lead to
panics that can't be handled as a user of the library. We bump the dependency
to master, which seems to fix those, but which also started adopting Futures
for the APIs we use.

To deal with this we refactor how the parallelization between getting the
current price and generating the rest of the position works, making it more
readable and clearer in the process.

One last detail: when building from git, yahoo-finance-rs relies on the protoc
binary existing on the system, so I moved to my own github fork for now. I
submitted a PR which shouold make it possible to go back to upstream's
repo: https://github.com/fbriden/yahoo-finance-rs/pull/20