Save weekly snapshots of Position while calculating
This will allow us to save our work when recalculating a position so that
we don't need to do everything from scratch every time, starting from the
last Position we have on file instead.

It also provides a very nice foundation in terms of data points for future
graphing features.