Add Event as a more general way of tracking asset timelines
Before this we were using operations. They are not a good way of representing
the timeline because there are several other kinds of events that could change
the portfolio other than purchases and sales.

Consider for instance stock splits, bonus shares, merging of stocks into units,
and so on. With this change we have a generic events collection with a detail
field which can be any of those. It also gives us a nice way of being
generic over asset types going forward.

To convert an existing operations collection use this:

db.getCollection('operations').aggregate([
    {
        "$project": {
            "symbol":  1,
            "time": 1,
            "eventType": "StockOperation",
            "detail": {
                "portfolios": "$portfolios",
                "assetType": "$assetType",
                "price": "$price",
                "quantity": "$quantity",
                "fees": "$fees",
                "type": "$type",
                "broker": "$broker"
            }
        }
    },
    {
        "$out": "events"
    }
])