Use temporary db with random name for tests
Tests were breaking when more than one was running in parallel (which
happened fairly often on my environment due to them being run by both
cargo watch and git push, for one). This was due to the fact that they
used a single database.

Now each test runs using its own temporary database with a generated
random name. It would be much better to use an in-memory storage, but
mongo seems to only make that available for the enterprise version (doh).

Also, beware panics, as they will not allow the database to be dropped.
I could not find a good way of fixing that yet, as panic::catch_unwind()
doesn't seem to work.