Add a PriceCache to speed up position endpoints
Position calculation was lagging a lot due to having to wait for a web
request to obtain the current price. Now we maintain a cache and keep
it updated using yahoo-finance-rs' streaming API.

This has lowered load time for Position endpoints on my local system
to ~10% of what it was.