Make tests more robust for github CI
I assumed initially that all tests ran as separate processes, which
is why they passed locally. Turns out they actually run as separate
threads on the same process, and that leads to a race.

If the position test ends in the middle of the historical one, or
vice-versa, the database that is being used by the test will get
dropped. Not a good thing to happen ;D

We now get each test to run on a separate process using rusty-fork,
and we also pass --nocapture to the test binary, so we can get any
println outputs as well.